feat(nest): rename features to api
