+++ assets/nest/example-app/test/hello-world.e2e-spec.ts
@@ -4,7 +4,7 @@ import {
   FastifyAdapter,
   NestFastifyApplication,
 } from '@nestjs/platform-fastify';
-import { HelloWorldModule } from '../src/features/hello-world/hello-world.module';
+import { HelloWorldModule } from '../src/api/hello-world/hello-world.module';
 
 describe('HelloWorldController (e2e)', () => {
   let app: NestFastifyApplication;
@@ -15,7 +15,7 @@ describe('HelloWorldController (e2e)', () => {
     }).compile();
 
     app = moduleFixture.createNestApplication<NestFastifyApplication>(
-      new FastifyAdapter(),
+      new FastifyAdapter()
     );
 
     await app.init();
